Change SVG dashboard editor layout
