--- app/scripts/components/svgEditor/edit/svgEdit.component.js
@@ -1,12 +1,0 @@
-import controller from './svgEdit.controller';
-import template from './svgEdit.html';
-
-export default {
-    restrict: 'E',
-    scope: true,
-    bindings: {
-        'dashboard': '='
-    },
-    template,
-    controller
-};
