+++ app/scripts/i18n/svg/ru.json
@@ -11,8 +11,8 @@
     "form": {
         "title": "Параметры панели",
         "field": {
-            "id": "Идентификатор",
-            "name": "Название",
+            "id": "Идентификатор панели",
+            "name": "Название панели",
             "fullwidth": "Растягивать во весь экран"
         },
         "error": {
@@ -104,12 +104,15 @@
             },
             "confirm": {
                 "remove": "Вы действительно хотите удалить эту привязку?"
+            },
+            "button": {
+                "remove": "Удалить"
             }
         },
         "params-json": {
-            "title" : "SVG привязки",
+            "title" : "Привязки",
             "button": {
-                "show" : "Редактировать в виде JSON"
+                "show" : "В виде JSON"
             }
         }
     },
@@ -118,7 +121,7 @@
         "save": "Сохранить",
         "cancel": "Отменить",
         "remove": "Удалить",
-        "view": "Показать панель",
+        "view": "Показать",
         "back": "Список панелей"
     },
