+++ app/styles/main.css
@@ -957,11 +957,6 @@ ul.col-md-2.nav.nav-pills.nav-stacked {
     margin-bottom: 9px;
 }
 
-.svg-edit-result > svg {
-    width: 100%;
-    height: 100%;
-}
-
 .modebar-btn .icon {
   width: 1.5em;
   height: 1.5em;
@@ -1090,8 +1085,7 @@ body.hmi .page-header {
 
 body.hmi #wrapper .navbar,
 body.hmi .page-header, 
-body.hmi .show-console, 
-body.hmi .svg-dashboard-footer {
+body.hmi .show-console {
   display: none
 }
 
@@ -1133,8 +1127,7 @@ body.hmi .svg-dashboard-footer {
   }
 
   body.fullscreen .access-level-label,
-  body.fullscreen .show-console,
-  body.fullscreen .svg-dashboard-footer {
+  body.fullscreen .show-console {
     display: none;
   }
 
@@ -1146,16 +1139,6 @@ body.hmi .svg-dashboard-footer {
     padding: 8px 23px 10px 23px;
     height: 100vh;
   }
-}
-
-body.fullscreen .page-header {
-  margin: 0px 0px 10px;
-}
-
-body.fullscreen .access-level-label,
-body.fullscreen .show-console,
-body.fullscreen .svg-dashboard-footer {
-  display: none;
 }
 
 .warning-notice {
@@ -1186,6 +1169,7 @@ body.fullscreen .svg-dashboard-footer {
   height: 100%;
 }
 
+.svg-edit-page,
 .svg-view-page {
   display: flex;
   flex-direction: column;
@@ -1205,4 +1189,8 @@ body.fullscreen .svg-dashboard-footer {
 .svg-view-fullscreen .svg-view-container > svg {
   height: 100%;
   width: 100%;
+}
+
+.svg-edit-page h1 {
+  margin-bottom: 7px;
 }
