+++ debian/changelog
@@ -1,3 +1,9 @@
+wb-mqtt-homeui (2.64.0) stable; urgency=medium
+
+  * Change SVG dashboard editor layout
+
+ -- Petr Krasnoshchekov <petr.krasnoshchekov@wirenboard.com>  Thu, 25 May 2023 09:42:17 +0500
+
 wb-mqtt-homeui (2.63.9) stable; urgency=medium
 
   * Fix duplicating values in widgets when corresponding service is restarted
